fixes to the dynamoDB code after testing on AWS

+++ web/src/main/kotlin/com/gmail/pzalejko/invoice/invoicerequest/model/InvoiceRequestRepository.kt
@@ -8,5 +8,5 @@ interface InvoiceRequestRepository {
 
     fun save(request: InvoiceRequest)
 
-    fun findLast(accountId: Int, month: Int, year: Int): InvoiceRequest?
+    fun findLast(accountId: Long, month: Int, year: Int): InvoiceRequest?
 }
